+++ central/role/datastore/datastore_test.go
@@ -27,6 +27,11 @@ func TestAllDefaultRolesAreCovered(t *testing.T) {
 	for r, a := range vulnRiskManagementDefaultRoles {
 		defaultRoles[r] = a
 	}
+	// Merge the roles for vuln reporting into the defaults
+	// TODO: Remove once the feature is released
+	for r, a := range vulnReportingDefaultRoles {
+		defaultRoles[r] = a
+	}
 
 	assert.Len(t, defaultRoles, len(role.DefaultRoleNames))
 	for r := range defaultRoles {